Add PlayerContext data class
