--- src/main/kotlin/ServerContext.kt
@@ -1,21 +1,0 @@
-package dev.deadzone
-
-import dev.deadzone.core.auth.AuthProvider
-import dev.deadzone.core.auth.SessionManager
-import dev.deadzone.data.db.BigDB
-import dev.deadzone.socket.OnlinePlayerRegistry
-
-data class ServerContext(
-    val db: BigDB,
-    val sessionManager: SessionManager,
-    val onlinePlayerRegistry: OnlinePlayerRegistry,
-    val authProvider: AuthProvider,
-    val config: ServerConfig,
-)
-
-data class ServerConfig(
-    val adminEnabled: Boolean,
-    val useMongo: Boolean,
-    val mongoUrl: String,
-    val isProd: Boolean,
-)
